fix: silence warnings pointed in issue #11917, minor performance fixes

+++ libraries/WebServer/src/Parsing.cpp
@@ -189,8 +189,8 @@ bool WebServer::_parseRequest(NetworkClient &client) {
       _currentHandler->raw(*this, _currentUri, *_currentRaw);
       _currentRaw->status = RAW_WRITE;
 
-      while (_currentRaw->totalSize < _clientContentLength) {
-        size_t read_len = std::min(_clientContentLength - _currentRaw->totalSize, (size_t)HTTP_RAW_BUFLEN);
+      while (_currentRaw->totalSize < (size_t)_clientContentLength) {
+        size_t read_len = std::min((size_t)_clientContentLength - _currentRaw->totalSize, (size_t)HTTP_RAW_BUFLEN);
         _currentRaw->currentSize = client.readBytes(_currentRaw->buf, read_len);
         _currentRaw->totalSize += _currentRaw->currentSize;
         if (_currentRaw->currentSize == 0) {
@@ -205,8 +205,8 @@ bool WebServer::_parseRequest(NetworkClient &client) {
       log_v("Finish Raw");
     } else if (!isForm) {
       size_t plainLength;
-      char *plainBuf = readBytesWithTimeout(client, _clientContentLength, plainLength, HTTP_MAX_POST_WAIT);
-      if (plainLength < _clientContentLength) {
+  char *plainBuf = readBytesWithTimeout(client, _clientContentLength, plainLength, HTTP_MAX_POST_WAIT);
+  if (plainLength < (size_t)_clientContentLength) {
         free(plainBuf);
         return false;
       }
@@ -407,7 +407,7 @@ int WebServer::_uploadReadByte(NetworkClient &client) {
 
 bool WebServer::_parseForm(NetworkClient &client, const String &boundary, uint32_t len) {
   (void)len;
-  log_v("Parse Form: Boundary: %s Length: %d", boundary.c_str(), len);
+  log_v("Parse Form: Boundary: %s Length: %u", boundary.c_str(), len);
   String line;
   int retry = 0;
   do {
@@ -432,7 +432,7 @@ bool WebServer::_parseForm(NetworkClient &client, const String &boundary, uint32
 
       line = client.readStringUntil('\r');
       client.readStringUntil('\n');
-      if (line.length() > 19 && line.substring(0, 19).equalsIgnoreCase(F("Content-Disposition"))) {
+  if (line.length() > (size_t)19 && line.substring(0, 19).equalsIgnoreCase(F("Content-Disposition"))) {
         int nameStart = line.indexOf('=');
         if (nameStart != -1) {
           argName = line.substring(nameStart + 2);
@@ -455,7 +455,7 @@ bool WebServer::_parseForm(NetworkClient &client, const String &boundary, uint32
           line = client.readStringUntil('\r');
           client.readStringUntil('\n');
           while (line.length() > 0) {
-            if (line.length() > 12 && line.substring(0, 12).equalsIgnoreCase(FPSTR(Content_Type))) {
+            if (line.length() > (size_t)12 && line.substring(0, 12).equalsIgnoreCase(FPSTR(Content_Type))) {
               argType = line.substring(line.indexOf(':') + 2);
             }
             //skip over any other headers
@@ -470,7 +470,7 @@ bool WebServer::_parseForm(NetworkClient &client, const String &boundary, uint32
               if (line.startsWith("--" + boundary)) {
                 break;
               }
-              if (argValue.length() > 0) {
+              if (argValue.length() > (size_t)0) {
                 argValue += "\n";
               }
               argValue += line;
